Separated out shared tag commands into a util file and moved invalid name error under a general tags section
