+++ modules/cli/pkg/tags/tagsSet.go
@@ -12,13 +12,11 @@ import (
 	"log"
 	"math"
 	"net/http"
-	"strings"
 
 	"github.com/galasa-dev/cli/pkg/embedded"
 	galasaErrors "github.com/galasa-dev/cli/pkg/errors"
 	"github.com/galasa-dev/cli/pkg/galasaapi"
 	"github.com/galasa-dev/cli/pkg/spi"
-	"github.com/galasa-dev/cli/pkg/utils"
 )
 
 const (
@@ -108,37 +106,4 @@ func sendTagUpdateToRestApi(
 		}
 	}
 	return tagGotBack, err
-}
-
-func validateTagName(tagName string) error {
-	var err error
-	log.Println("Validating the provided tag name")
-
-	err = validateStringIsLatin1AndNotBlank(tagName, galasaErrors.GALASA_ERROR_SET_TAG_INVALID_NAME)
-
-	if err == nil {
-		log.Println("Tag name validated OK")
-	}
-	return err
-}
-
-func validateDescription(description string) error {
-	var err error
-	log.Println("Validating the provided description")
-
-	err = validateStringIsLatin1AndNotBlank(description, galasaErrors.GALASA_ERROR_INVALID_TAG_DESCRIPTION)
-	if err == nil {
-		log.Println("Description validated OK")
-	}
-	return err
-}
-
-func validateStringIsLatin1AndNotBlank(str string, errMessageType *galasaErrors.MessageType) error {
-    var err error
-    str = strings.TrimSpace(str)
-
-    if str == "" || !utils.IsLatin1(str) {
-        err = galasaErrors.NewGalasaError(errMessageType)
-    }
-    return err
 }
